+++ bash/portal/logging.sh
@@ -36,23 +36,7 @@
 PROJECTROOTDIRECTORY="$PWD"
 BUILDDIRECTORY="$PROJECTROOTDIRECTORY/build"
 PORTALBUILDDIRECTORY="$BUILDDIRECTORY/portal"
-
-# Assign the Lighthttpd document root directory to a variable.
-RAWDOCUMENTROOT=`/usr/sbin/lighttpd -f /etc/lighttpd/lighttpd.conf -p | grep server.document-root`
-LIGHTTPDDOCUMENTROOT=`sed 's/.*"\(.*\)"[^"]*$/\1/' <<< $RAWDOCUMENTROOT`
-
-DATABASEENGINE=$ADSB_DATABASEENGINE
-DATABASEHOSTNAME=$ADSB_DATABASEHOSTNAME
-DATABASEUSER=$ADSB_DATABASEUSER
-DATABASEPASSWORD1=$ADSB_DATABASEPASSWORD1
-DATABASENAME=$ADSB_DATABASENAME
-
-unset ADSB_DATABASEENGINE
-unset ADSB_DATABASEHOSTNAME
-unset ADSB_DATABASEUSER
-unset ADSB_DATABASEPASSWORD1
-unset ADSB_DATABASENAME
-
+PORTALPYTHONDIRECTORY="$PORTALBUILDDIRECTORY/python"
 PYTHONPATH=`which python`
 
 ## SETUP FLIGHT LOGGING
@@ -61,84 +45,50 @@ echo ""
 echo -e "\e[95m  Setting up flight logging...\e[97m"
 echo ""
 
-case $DATABASEENGINE in
-    "MySQL")
-        echo -e "\e[94m  Creating the flight logging configuration file for MySQL...\e[97m"
-        tee $PORTALBUILDDIRECTORY/logging/config.json > /dev/null <<EOF
-{
-    "database":{"type":"mysql",
-                "host":"$DATABASEHOSTNAME",
-                "user":"$DATABASEUSER",
-                "passwd":"$DATABASEPASSWORD1",
-                "db":"$DATABASENAME"}
-}
-EOF
-            ;;
-    "SQLite")
-        echo -e "\e[94m  Creating the flight logging configuration file for SQLite...\e[97m"
-        tee $PORTALBUILDDIRECTORY/logging/config.json > /dev/null <<EOF
-{
-    "database":{"type":"sqlite",
-                "host":"$DATABASEHOSTNAME",
-                "user":"$DATABASEUSER",
-                "passwd":"$DATABASEPASSWORD1",
-                "db":"$DATABASENAME"}
-}
-EOF
-        ;;
-    *)
-        echo ""
-        echo -e "\e[91m  \e[5mINSTALLATION HALTED!\e[25m"
-        echo -e "  SETUP HAS BEEN TERMINATED!"
-        echo ""
-        echo -e "\e[93mInvalid \"DATABASEENGINE\" supplied.\e[39m"
-        echo ""
-        echo -e "\e[93m----------------------------------------------------------------------------------------------------"
-        echo -e "\e[92m  ADS-B Receiver Project Portal (Advanced) setup.\e[39m"
-        echo ""
-        read -p "Press enter to continue..." CONTINUE
-        exit 1
-        ;;
-esac
-
 # Create and set permissions on the flight logging and maintenance maintenance scripts.
 echo -e "\e[94m  Creating the flight logging maintenance script...\e[97m"
-tee $PORTALBUILDDIRECTORY/logging/flights-maint.sh > /dev/null <<EOF
+tee $PORTALPYTHONDIRECTORY/flights-maint.sh > /dev/null <<EOF
 #!/bin/sh
 while true
   do
     sleep 30
-        $PYTHONPATH $PORTALBUILDDIRECTORY/logging/flights.py
+        $PYTHONPATH $PORTALPYTHONDIRECTORY/flights.py
   done
 EOF
 
 echo -e "\e[94m  Creating the maintenance maintenance script...\e[97m"
-tee $PORTALBUILDDIRECTORY/logging/maintenance-maint.sh > /dev/null <<EOF
+tee $PORTALPYTHONDIRECTORY/maintenance-maint.sh > /dev/null <<EOF
 #!/bin/sh
 while true
   do
     sleep 30
-        $PYTHONPATH $PORTALBUILDDIRECTORY/logging/maintenance.py
+        $PYTHONPATH $PORTALPYTHONDIRECTORY/maintenance.py
   done
 EOF
 
 echo -e "\e[94m  Making the flight logging maintenance script executable...\e[97m"
-chmod +x $PORTALBUILDDIRECTORY/logging/flights-maint.sh
+chmod +x $PORTALPYTHONDIRECTORY/flights-maint.sh
 echo -e "\e[94m  Making the maintenance maintenance script executable...\e[97m"
-chmod +x $PORTALBUILDDIRECTORY/logging/maintenance-maint.sh
+chmod +x $PORTALPYTHONDIRECTORY/maintenance-maint.sh
+
+#Remove old flights-maint.sh start up line from /etc/rc.local.
+sudo sed -i '/build\/portal\/logging\/flights-maint.sh/d' /etc/rc.local
 
 # Add flight logging maintenance script to rc.local.
-if ! grep -Fxq "$PORTALBUILDDIRECTORY/logging/flights-maint.sh &" /etc/rc.local; then
+if ! grep -Fxq "$PORTALPYTHONDIRECTORY/flights-maint.sh &" /etc/rc.local; then
     echo -e "\e[94m  Adding the flight logging maintenance script startup line to /etc/rc.local...\e[97m"
     LINENUMBER=($(sed -n '/exit 0/=' /etc/rc.local))
-    ((LINENUMBER>0)) && sudo sed -i "${LINENUMBER[$((${#LINENUMBER[@]}-1))]}i $PORTALBUILDDIRECTORY/logging/flights-maint.sh &\n" /etc/rc.local
+    ((LINENUMBER>0)) && sudo sed -i "${LINENUMBER[$((${#LINENUMBER[@]}-1))]}i $PORTALPYTHONDIRECTORY/flights-maint.sh &\n" /etc/rc.local
 fi
 
+# Remove old maintenance-maint.sh start up line from /etc/rc.local.
+sudo sed -i '/build\/portal\/logging\/maintenance-maint.sh/d' /etc/rc.local
+
 # Add maintenance maintenance script to rc.local.
-if ! grep -Fxq "$PORTALBUILDDIRECTORY/logging/maintenance-maint.sh &" /etc/rc.local; then
+if ! grep -Fxq "$PORTALPYTHONDIRECTORY/maintenance-maint.sh &" /etc/rc.local; then
     echo -e "\e[94m  Adding the maintenance maintenance script startup line to /etc/rc.local...\e[97m"
     LINENUMBER=($(sed -n '/exit 0/=' /etc/rc.local))
-    ((LINENUMBER>0)) && sudo sed -i "${LINENUMBER[$((${#LINENUMBER[@]}-1))]}i $PORTALBUILDDIRECTORY/logging/maintenance-maint.sh &\n" /etc/rc.local
+    ((LINENUMBER>0)) && sudo sed -i "${LINENUMBER[$((${#LINENUMBER[@]}-1))]}i $PORTALPYTHONDIRECTORY/maintenance-maint.sh &\n" /etc/rc.local
 fi
 
 # Kill any previously running maintenance scripts.
@@ -158,10 +108,10 @@ fi
 
 # Start flight logging.
 echo -e "\e[94m  Executing the flight logging maintenance script...\e[97m"
-nohup $PORTALBUILDDIRECTORY/logging/flights-maint.sh > /dev/null 2>&1 &
+nohup $PORTALPYTHONDIRECTORY/flights-maint.sh > /dev/null 2>&1 &
 
 # Start maintenance.
 echo -e "\e[94m  Executing the maintenance maintenance script...\e[97m"
-nohup $PORTALBUILDDIRECTORY/logging/maintenance-maint.sh > /dev/null 2>&1 &
+nohup $PORTALPYTHONDIRECTORY/maintenance-maint.sh > /dev/null 2>&1 &
 
 exit 0